+++ lib/src/users/patron/patron_user_component.dart
@@ -1,0 +1,2 @@
+library plato.cap.components.user.patron;
+